refactor: extract contacts CLI command into its own module
